Uses an auxiliary script for token insertion.
In the previous implementation, the JSON claims object was stored as a
string value and later delivered to the `docker exec ... psql` through
the shell.
Since Windows and Unix shells handle escaping of characters
differently, a Unix shell script was written for use inside the
container. This way, the JSON string is created inside the container.