Use dedicated docker image for Token Insertion.
Previously, we were mounting a directory from the host to the container
and running a script from that directory.
However, the differences between Win vs UNIX in file formatting
(specifically \r\n vs \n) would end up in error if the container
(running against a Linux kernel) attempted to run the script located
in the Win host.